+++ src/internal/errors.go
@@ -1,0 +1,7 @@
+package internal
+
+import "net/http"
+
+var (
+	ErrUserNotFound = NewHttpError(http.StatusNotFound, "user not found")
+)
